feat: Implemented Manager to asynchronously handle both TelegramBot and TradingBot
